WebCLMatcher1Action::checkIdentifiers: fixed a compiler warning
Removed a warning about size_t converted implicitly to const int.

+++ src/WebCLAction.cpp
@@ -227,7 +227,7 @@ bool WebCLMatcher1Action::checkIdentifiers()
         cfg_.variablePrefix_.c_str(),
         cfg_.macroPrefix_.c_str()
     };
-    const int lengths[numPrefixes] = {
+    const size_t lengths[numPrefixes] = {
         cfg_.typePrefix_.size(),
         cfg_.variablePrefix_.size(),
         cfg_.macroPrefix_.size()
